+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - Fixed rendering `{{ block.super }}` with several levels of inheritance
 - Fixed checking dictionary values for nil in `default` filter
 - Fixed comparing string variables with string literals, in Swift 4 string literals became `Substring` and thus couldn't be directly compared to strings.
-
+- Fixed using spaces between loop variables
 
 ## 0.10.1